requirements.txt
streamlit
transformers
torch
pillow
sqlalchemy<2.0.0
pandas
plotly